Finished millisecond timers

+++ include/timing/timing.h
@@ -11,7 +11,7 @@ void timer0_init_as_timer_ms(float ms, void (*callback)(void));
 void timer1_init_as_pwm(void);
 void timer1_set_duty_cycle_a(float duty_cycle);
 void timer1_set_duty_cycle_b(float duty_cycle);
-void timer1_init_as_timer_ms(uint32_t ms, void (*callback)(void));
+void timer1_init_as_timer_ms(float ms, void (*callback)(void));
 
 void timer2_init_as_pwm(void);
 void timer2_set_duty_cycle_a(float duty_cycle);
